W173i: h_pad 00103DA0 (H_Pad_RwAllocateRaw) 424/432 nd215 -> 420/432 MATCH by recovering the CHECK string's GP-RELATIVE address mode with an .sdata-qualified alias. h_pad 00104140 (H_Pad_RwMalloc) 312/320 nd147 -> 316/320 nd4 using the same alias plus a two-value OR reconstructed as an ascending-label switch. h_cursor 00100230 nd97 -> 67 via base-first global and typed indexed accesses. k_command 001C5980 828/944 -> 924/944, 96 bytes recovered toward its window via retail's ascending-case switch, separate call guards, candidate-unit traversal and per-case array-base roles.
mdlFile 0031E420 172/176 nd50 -> MATCH and 00324510 168/160 -> 148/160, window cleared; 0031E310 recovered 60 bytes of missing retail work, cutting its deficit from 68 to 8.

The GP-relative-vs-absolute string address mode is a new member of the addressing-mode defect family alongside the _abs data aliases -- worth checking wherever a CHECK/assert string is referenced.

+++ src/Main/Battle/Cmd/bsa_main.c
@@ -26,7 +26,7 @@ extern code D_0096009C;
 
 extern void* func_001ff430(u32 id);
 extern void func_0017b1e0(u16 id);
-extern s16 func_0017d2e0(void);
+extern s32 func_0017d2e0(void);
 extern u32 func_0010a4e0(s32 a, s32 b, s32 c, s32 d);
 
 extern u32 func_0021c3f0(s32 texture);
@@ -275,7 +275,7 @@ void bsaMain0020fe30(u32* p, s32 mode, u32 unitId)
         func_0021e380(p + i * 0x80 + 0xc90, image, 1);
     }
     for (i = 0; i < 8; i++) {
-        if (*(s16*)(unit + i * 2) == 0)
+        if (*(u16*)(unit + i * 2) == 0)
             *(s16*)((u8*)p + i * 2 + 0xa6b8) = 0x2b;
         else
             *(s16*)((u8*)p + i * 2 + 0xa6b8) = func_0017d2e0() + 0x20;
